RAG system/pyproject.toml
[project]
name = "ragdemo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "faiss-cpu>=1.12.0",
    "ipykernel>=7.1.0",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-openai>=0.3.30",
    "langgraph>=0.6.5",
    "openai>=1.99.9",
    "openpyxl>=3.1.5",
    "pandas>=2.3.1",
    "pydantic>=2.11.7",
    "pypdf>=6.0.0",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "scikit-learn>=1.7.2",
    "sentence-transformers>=5.1.2",
    "streamlit>=1.48.1",
    "tqdm>=4.67.1",
    "wikipedia>=1.4.0",
]
